+++ AlphaFS/Filesystem/FileSystemInfo.cs
@@ -169,7 +169,7 @@ namespace Alphaleonis.Win32.Filesystem
          [SecurityCritical]
          set
          {
-            File.SetFsoDateTimeCore(Transaction, LongFullName, value, null, null, false, PathFormat.LongFullPath);
+            File.SetFsoDateTimeCore(Transaction, IsDirectory, LongFullName, value, null, null, false, PathFormat.LongFullPath);
             Reset();
          }
       }
@@ -270,7 +270,7 @@ namespace Alphaleonis.Win32.Filesystem
          [SecurityCritical]
          set
          {
-            File.SetFsoDateTimeCore(Transaction, LongFullName, null, value, null, false, PathFormat.LongFullPath);
+            File.SetFsoDateTimeCore(Transaction, IsDirectory, LongFullName, null, value, null, false, PathFormat.LongFullPath);
             Reset();
          }
       }
@@ -326,7 +326,7 @@ namespace Alphaleonis.Win32.Filesystem
          [SecurityCritical]
          set
          {
-            File.SetFsoDateTimeCore(Transaction, LongFullName, null, null, value, false, PathFormat.LongFullPath);
+            File.SetFsoDateTimeCore(Transaction, IsDirectory, LongFullName, null, null, value, false, PathFormat.LongFullPath);
             Reset();
          }
       }