+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+Version 2.2.4  (2018-XX-XX)
+-------------
+
+### Bugs Fixed
+
+- Issue #468: Method `Directory.CopyTimestamps` should throw `DirectoryNotFoundException`.
+- Issue #469: Method `Directory.GetFileIdInfo` should throw `DirectoryNotFoundException`.
+
+
 Version 2.2.3  (2018-06-14)
 -------------